알고리즘 2438 문제풀이

+++ 222018/1008.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-func divide(A: Decimal, B: Decimal) {
+func divide(A: Double, B: Double) {
     let divide = A/B
     let numberFormater = NumberFormatter()
     numberFormater.maximumFractionDigits = 9
